Render #26: the terrain grid that swallowed every ground-flush decal
Unit decals showed on some soldiers and not others, flickered as the
camera moved, and left only a sliver of colour at the feet. Nothing was
wrong with the decals.

SDLTileMapRenderer picked its mesh step so the grid fit a 16-bit index
buffer -- while(((H/step)+1)*((V/step)+1) >= 65536) step *= 2 -- which on
Maelstrom's 2048x4096 heightfield is step 16: one grid vertex per 16 world
units, a 129x257 grid. Everything that lies flat on the ground is placed
against the true per-cell heightfield with almost no clearance, so where
the grid interpolated above the real surface it took the lot with it:

  - CircleManager::drawSpline's select ring, at getAltWhole + 1, one unit
  - the flat plate at the base of a unit's own 3DX model -- the red
    triangle on a soldier, the clan-symbol emblem on a commander. Same
    plate, different UVs: it takes the legion skin colour, and
    SetSkinColor(unitColor(), unitSign()) composites the emblem in at
    logoPosition_, which is why the shape varies per unit type and the
    colour never does.

The cap was addressing, not memory, so: 32-bit index buffer
(buildIndexData emits unsigned int, both SDL_BindGPUIndexBuffer sites take
INDEXELEMENTSIZE_32BIT, both upload paths size against it), and the step
driven by a 600k-vertex budget that also searches downward from STEP_BASE
-- so small maps get finer than before rather than merely not-coarser.
512 now draws at step 1, 1024 at 2, 2048x4096 at 4.

A slope-scaled depth bias (factor 2, constant 0 -- its unit is
implementation-defined against D32_FLOAT) covers the residual, on the
lava and ice pipelines as well as the terrain's, since those replace the
terrain pipeline over the same mesh and a decal on lava would otherwise
sink where one on soil does not.

This is accuracy bought uniformly, by memory; the original bought it where
it mattered, with per-tile LOD. Register #26 says to restore the LOD if
the budget ever binds rather than raise it further.

+++ Render/SDLTileMapRenderer.h
@@ -111,7 +111,7 @@ private:
 	// One grid vertex from vMap: world position + normal, as buildMesh samples them.
 	void computeVertex(Vertex& v, int gx, int gy) const;
 	// quadMat_ -> index buffer grouped into one contiguous run per material (fills runs_).
-	void buildIndexData(std::vector<unsigned short>& idx);
+	void buildIndexData(std::vector<unsigned int>& idx);
 	// Terramorphing: consume the ATTRTILE_UPDATE_* flags cTileMap::BuildRegionPoint raised
 	// for vMap's update rects, recompute the covered mesh rows and colour texels in the CPU
 	// mirrors and upload just those; rebuild the index runs if a cell's material repainted.
